Add example manifest for breath counting
My first foray into what these manifests will look like, where the line
between the server, parameters, and client sides of things will go.

Notably: If we rely on the instrument to give us values in our database,
we can do a whole lot of things.

It might be sane to be able to add an array of tags to measurements?
I don't know about that yet, or how it would affect performance...